requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
qrcode[pil]==7.4.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
